+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 # 3.3.2
 ```
     * In AbstractRequest class, getParams() method, was added additional check for some of the merchant details.
-    * In ConfigProvider class, getWebSdkConfig() method we will typecast the "sdk_style" to string.
+    * In ConfigProvider class, will typecast the "sdk_style" to string.
 ```
 
 # 3.3.1